backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.1.2
certifi==2025.8.3
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.111.0
fastapi-cli==0.0.10
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.11.3
psycopg2-binary==2.9.9
pyasn1==0.6.1
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.19.2
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.1.0
rich-toolkit==0.15.1
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
typer==0.17.4
typing_extensions==4.15.0
ujson==5.11.0
uvicorn==0.29.0
watchfiles==1.1.0
websockets==15.0.1
